% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{air_polution_mortality_df}
\alias{air_polution_mortality_df}
\title{Air Pollution and Mortality}
\format{
A data frame with 60 observations and 7 variables:
\describe{
  \item{City}{Metropolitan area (factor with 60 levels)}
  \item{Mort}{Mortality rate (numeric vector)}
  \item{Precip}{Annual precipitation in inches (integer vector)}
  \item{Educ}{Median years of education (numeric vector)}
  \item{NonWhite}{Percentage of non-white population (numeric vector)}
  \item{NOX}{Nitrogen oxide concentration (integer vector)}
  \item{SO2}{Sulfur dioxide concentration (integer vector)}
}
}
\source{
Data taken from the Sleuth3 package version 1.0-6
}
\usage{
data(air_polution_mortality_df)
}
\description{
This dataset, air_polution_mortality_df, is a data frame containing information from
an early study exploring the relationship between air pollution and mortality across
5 Standard Metropolitan Statistical Areas in the U.S. between 1959 and 1961.
}
\details{
The dataset name has been kept as 'air_polution_mortality_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
