% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{bronchitis_Cardiff_df}
\alias{bronchitis_Cardiff_df}
\title{Chronic Bronchitis in Cardiff Men}
\format{
A data frame with 212 observations and 4 variables:
\describe{
  \item{cig}{Number of cigarettes smoked per day (numeric)}
  \item{poll}{Pollution exposure level (numeric)}
  \item{r}{Bronchitis diagnosis (0 = no, 1 = yes) (integer)}
  \item{rfac}{Bronchitis diagnosis as a factor with 2 levels (factor)}
}
}
\source{
Data taken from the gamclass package version 0.62.5
}
\usage{
data(bronchitis_Cardiff_df)
}
\description{
This dataset, bronchitis_Cardiff_df, is a data frame containing information from a study
assessing the effects of smoking and pollution on bronchitis diagnosis in a sample of 212 men
from Cardiff.
}
\details{
The dataset name has been kept as 'bronchitis_Cardiff_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
