% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{children_respiratory_rates_df}
\alias{children_respiratory_rates_df}
\title{Children Respiratory Rates Data}
\format{
A data frame with 618 observations and 2 variables:
\describe{
  \item{Age}{Child's age in days (numeric vector)}
  \item{Rate}{Respiratory rate in breaths per minute (integer vector)}
}
}
\source{
Data taken from the Sleuth3 package version 1.0-6
}
\usage{
data(children_respiratory_rates_df)
}
\description{
This dataset, children_respiratory_rates_df, is a data frame containing respiratory rate
measurements from 618 Italian children aged between 15 days and 3 years, collected to
establish normal respiratory rate distributions for clinical assessment.
}
\details{
The dataset name has been kept as 'children_respiratory_rates_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
