% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{danish_lung_incidence_df}
\alias{danish_lung_incidence_df}
\title{Lung cancer in 4 Danish cities 1968-71}
\format{
A data frame with 24 observations and 4 variables:
\describe{
  \item{city}{City of observation (factor with 4 levels)}
  \item{age}{Age group (factor with 6 levels)}
  \item{pop}{Population size (integer)}
  \item{cases}{Number of incident lung cancer cases (integer)}
}
}
\source{
Data taken from the ISwR package version 2.0-10
}
\usage{
data(danish_lung_incidence_df)
}
\description{
This dataset, danish_lung_incidence_df, is a data frame containing counts of
incident lung cancer cases and population size in four neighbouring Danish cities
by age group from 1968 to 1971.
}
\details{
The dataset name has been kept as 'danish_lung_incidence_df' to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the PulmoDataSets package. The suffix 'df' indicates
that this is a data frame object. The original content has not been modified
in any way.
}
