% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{influenza_us_1975_df}
\alias{influenza_us_1975_df}
\title{US 1975-76 Influenza-Like Illness Data}
\format{
A data frame with 49 observations (states + DC) and 7 variables:
\describe{
  \item{State}{State identifier (integer)}
  \item{Acronym}{State abbreviation (factor with 51 levels)}
  \item{Pop}{State population (integer)}
  \item{Latitude}{Geographic latitude (numeric)}
  \item{Longitude}{Geographic longitude (numeric)}
  \item{Start}{Week of season start (integer)}
  \item{Peak}{Week of peak activity (integer)}
}
}
\source{
Data taken from the epimdr package version 0.6-5
}
\usage{
data(influenza_us_1975_df)
}
\description{
This dataset, influenza_us_1975_df, is a data frame containing influenza-like illness (ILI)
data for the lower 48 US states and District of Columbia during the 1975-76 season,
which was dominated by the A H3N2 Victoria strain.
}
\details{
The dataset name has been kept as 'influenza_us_1975_df' to avoid confusion with other
datasets in the R ecosystem. This naming convention helps distinguish this dataset
as part of the PulmoDataSets package. The suffix 'df' indicates that this is a
standard data frame. The original content has not been modified in any way.
}
