% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ohio_children_wheeze_df}
\alias{ohio_children_wheeze_df}
\title{Ohio Children Wheeze Status}
\format{
A data frame with 2148 observations and 4 variables:
\describe{
  \item{resp}{Wheeze status (0 = no wheeze, 1 = wheeze) (integer)}
  \item{id}{Child identifier (integer)}
  \item{age}{Age of the child in years (integer)}
  \item{smoke}{Parental smoking status (0 = no, 1 = yes) (integer)}
}
}
\source{
Data taken from the geepack package version 1.3.12
}
\usage{
data(ohio_children_wheeze_df)
}
\description{
This dataset, ohio_children_wheeze_df, is a data frame containing wheeze status data
from 2148 observations of children in Ohio. The data are part of a subset from the Six-City Study,
a longitudinal study examining the health effects of air pollution on children.
}
\details{
The dataset name has been kept as 'ohio_children_wheeze_df' to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the PulmoDataSets package. The suffix 'df' indicates
that this is a data frame object. The original content has not been modified
in any way.
}
