% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{pneumonia_influenza_ts}
\alias{pneumonia_influenza_ts}
\title{Monthly Pneumonia and Influenza Deaths in the U.S.}
\format{
A time series with 132 monthly observations from January 1968 to December 1978:
\describe{
  \item{Value}{Mortality rate (numeric vector)}
  \item{Time}{Monthly index from 1968 to 1978 (time series vector)}
}
}
\source{
Data taken from the astsa package version 2.2
}
\usage{
data(pneumonia_influenza_ts)
}
\description{
This dataset, pneumonia_influenza_ts, is a time series containing monthly rates
of pneumonia and influenza deaths in the United States from 1968 to 1978.
}
\details{
The dataset name has been kept as 'pneumonia_influenza_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series. The original content has not been modified
in any way.
}
