% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmodatasets-package.R
\docType{package}
\name{PulmoDataSets}
\alias{PulmoDataSets}
\alias{PulmoDataSets-package}
\title{PulmoDataSets: A Curated Collection of Pulmonary and Respiratory Disease Datasets}
\description{
{
This package provides a wide variety of datasets focused on the lungs, respiratory system, tuberculosis, whooping cough, pneumonia, influenza and associated diseases.
}
}
\details{
PulmoDataSets: A Curated Collection of Pulmonary and Respiratory Disease Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Curated Collection of Pulmonary and Respiratory Disease Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/pulmodatasets}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
