% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{respiratory_trial_outcomes_df}
\alias{respiratory_trial_outcomes_df}
\title{Ordinal respiratory outcomes}
\format{
A data frame with 111 observations and 5 variables:
\describe{
  \item{y1}{Ordinal response at visit 1 (integer)}
  \item{y2}{Ordinal response at visit 2 (integer)}
  \item{y3}{Ordinal response at visit 3 (integer)}
  \item{y4}{Ordinal response at visit 4 (integer)}
  \item{trt}{Treatment group (0 = placebo, 1 = active) (integer)}
}
}
\source{
Data taken from the geepack package version 1.3.12
}
\usage{
data(respiratory_trial_outcomes_df)
}
\description{
This dataset, respiratory_trial_outcomes_df, is a data frame containing outcome data
from a randomized clinical trial described in Miller et al. (1993) evaluating a new
treatment for respiratory disorder. The study includes 111 patients who were randomly assigned
to one of two treatments (active or placebo). The patients were followed up at four visits,
and their response status was classified on an ordinal scale at each visit.
}
\details{
The dataset name has been kept as 'respiratory_trial_outcomes_df' to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the PulmoDataSets package. The suffix 'df' indicates
that this is a data frame object. The original content has not been modified
in any way.
}
