% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{smoking_lung_cancer_df}
\alias{smoking_lung_cancer_df}
\title{Smoking and Lung Cancer}
\format{
A data frame with 2 observations and 3 variables:
\describe{
  \item{Smoking}{Smoking status (factor with 2 levels: "NonSmokers", "Smokers")}
  \item{Cancer}{Number of lung cancer cases (integer vector)}
  \item{Control}{Number of control cases (integer vector)}
}
}
\source{
Data taken from the Sleuth3 package version 1.0-6
}
\usage{
data(smoking_lung_cancer_df)
}
\description{
This dataset, smoking_lung_cancer_df, is a data frame containing data from a retrospective
case-control study comparing smoking status between 86 lung cancer patients and 86 controls.
}
\details{
The dataset name has been kept as 'smoking_lung_cancer_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
