% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{smoking_UK_tbl_df}
\alias{smoking_UK_tbl_df}
\title{UK Smoking Habits}
\format{
A tibble with 1,691 observations and 12 variables:
\describe{
  \item{gender}{Gender of respondent (factor with 2 levels)}
  \item{age}{Age in years (integer)}
  \item{marital_status}{Marital status (factor with 5 levels)}
  \item{highest_qualification}{Highest education qualification (factor with 8 levels)}
  \item{nationality}{Nationality (factor with 8 levels)}
  \item{ethnicity}{Ethnic group (factor with 7 levels)}
  \item{gross_income}{Income bracket (factor with 10 levels)}
  \item{region}{UK region (factor with 7 levels)}
  \item{smoke}{Smoking status (factor with 2 levels)}
  \item{amt_weekends}{Cigarettes smoked on weekends (integer)}
  \item{amt_weekdays}{Cigarettes smoked on weekdays (integer)}
  \item{type}{Type of tobacco used (factor with 5 levels)}
}
}
\source{
Data taken from the openintro package version 2.5.0
}
\usage{
data(smoking_UK_tbl_df)
}
\description{
This dataset, smoking_UK_tbl_df, is a tibble containing survey data on smoking habits
from the UK, with demographic characteristics and tobacco consumption patterns
from 1,691 respondents.
}
\details{
The dataset name has been kept as 'smoking_UK_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package. The suffix 'tbl_df' indicates that this is a tibble data frame.
The original content has not been modified in any way.
}
