% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{smoking_youth_tbl_df}
\alias{smoking_youth_tbl_df}
\title{Youth Smoking and Lung Function}
\format{
A tibble with 654 observations and 5 variables:
\describe{
  \item{age}{Age in years (integer)}
  \item{FEV}{Forced Expiratory Volume in liters (numeric)}
  \item{height}{Height in centimeters (numeric)}
  \item{sex}{Sex of participant (character)}
  \item{smoker}{Smoking status (character)}
}
}
\source{
Data taken from the LSTbook package version 0.6
}
\usage{
data(smoking_youth_tbl_df)
}
\description{
This dataset, smoking_youth_tbl_df, is a tibble containing data from the Childhood
Respiratory Disease Study collected in the late 1970s, examining the effects of
smoking and second-hand smoke exposure on pulmonary function in 654 youths.
}
\details{
The dataset name has been kept as 'smoking_youth_tbl_df' to avoid confusion with other
datasets in the R ecosystem. This naming convention helps distinguish this dataset
as part of the PulmoDataSets package. The suffix 'tbl_df' indicates that this is a
tibble data frame. The original content has not been modified in any way.
}
