% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{tlc_lung_capacity_df}
\alias{tlc_lung_capacity_df}
\title{Total Lung Capacity}
\format{
A data frame with 32 observations and 4 variables:
\describe{
  \item{age}{Age in years (integer)}
  \item{sex}{Sex (0 = female, 1 = male) (integer)}
  \item{height}{Height in centimeters (integer)}
  \item{tlc}{Total lung capacity in liters (numeric)}
}
}
\source{
Data taken from the ISwR package version 2.0-10
}
\usage{
data(tlc_lung_capacity_df)
}
\description{
This dataset, tlc_lung_capacity_df, is a data frame containing data on pretransplant total
lung capacity (TLC) measured by whole-body plethysmography for recipients of heart-lung transplants.
}
\details{
The dataset name has been kept as 'tlc_lung_capacity_df' to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the PulmoDataSets package. The suffix 'df' indicates
that this is a data frame object. The original content has not been modified
in any way.
}
