% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{tuberculosis_vaccine_df}
\alias{tuberculosis_vaccine_df}
\title{BCG Vaccine Against Tuberculosis}
\format{
A data frame with 13 observations and 9 variables:
\describe{
  \item{trial}{Trial identifier number (integer vector)}
  \item{author}{Study author(s) (character vector)}
  \item{year}{Year of publication (integer vector)}
  \item{tpos}{Number of TB positive cases in vaccinated group (integer vector)}
  \item{tneg}{Number of TB negative cases in vaccinated group (integer vector)}
  \item{cpos}{Number of TB positive cases in control group (integer vector)}
  \item{cneg}{Number of TB negative cases in control group (integer vector)}
  \item{ablat}{Absolute latitude of study location (integer vector)}
  \item{alloc}{Method of treatment allocation (character vector)}
}
}
\source{
Data taken from the metadat package version 1.4-0
}
\usage{
data(tuberculosis_vaccine_df)
}
\description{
This dataset, tuberculosis_vaccine_df, is a data frame containing results from 13 clinical trials
examining the effectiveness of the Bacillus Calmette-Guerin (BCG) vaccine against tuberculosis.
}
\details{
The dataset name has been kept as 'tuberculosis_vaccine_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
