% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{UK_male_lung_deaths_ts}
\alias{UK_male_lung_deaths_ts}
\title{UK Male Lung Disease Deaths}
\format{
A time series (ts) object with 72 monthly observations from 1974 to 1979.
\describe{
  \item{value}{Number of deaths (numeric vector)}
  \item{time}{Time index (1974 to 1979)}
}
}
\source{
Data taken from the datasets package (R version 4.5.0), mdeaths dataset
}
\usage{
data(UK_male_lung_deaths_ts)
}
\description{
This dataset, UK_male_lung_deaths_ts, is a time series object containing monthly deaths
from bronchitis, emphysema and asthma in the UK from 1974 to 1979, for males.
}
\details{
The dataset name has been kept as 'UK_male_lung_deaths_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series object. The original content has not been modified
in any way.
}
