% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USMortality_df}
\alias{USMortality_df}
\title{US Mortality Rates by Cause and Gender}
\format{
A data frame with 40 observations and 5 variables:
\describe{
  \item{Status}{Rural/Urban status (factor with 2 levels)}
  \item{Sex}{Gender (factor with 2 levels)}
  \item{Cause}{Cause of death (factor with 10 levels)}
  \item{Rate}{Mortality rate (numeric vector)}
  \item{SE}{Standard error of mortality rate (numeric vector)}
}
}
\source{
Data taken from the lattice package version 0.22-6
}
\usage{
data(USMortality_df)
}
\description{
This dataset, USMortality_df, is a data frame containing mortality rates across all ages
in the USA by cause of death, sex, rural and urban status from 2011 to 2013. The data
represent national aggregate rates under the Department of Health and Human Services (HHS).
}
\details{
The dataset name has been kept as 'USMortality_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a standard data frame. The original content has not been modified
in any way.
}
