% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USRegionalMortality_df}
\alias{USRegionalMortality_df}
\title{US Regional Mortality Rates by Cause and Gender}
\format{
A data frame with 400 observations and 6 variables:
\describe{
  \item{Region}{HHS administrative region (factor with 10 levels)}
  \item{Status}{Rural/Urban status (factor with 2 levels)}
  \item{Sex}{Gender (factor with 2 levels)}
  \item{Cause}{Cause of death (factor with 10 levels)}
  \item{Rate}{Mortality rate (numeric vector)}
  \item{SE}{Standard error of mortality rate (numeric vector)}
}
}
\source{
Data taken from the lattice package version 0.22-6
}
\usage{
data(USRegionalMortality_df)
}
\description{
This dataset, USRegionalMortality_df, is a data frame containing region-wise mortality rates
across all ages in the USA by cause of death, sex, rural and urban status from 2011 to 2013.
The data represent rates for each administrative region under the Department of Health and Human Services (HHS).
}
\details{
The dataset name has been kept as 'USRegionalMortality_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
PulmoDataSets package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
