% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{veterans_lung_cancer_df}
\alias{veterans_lung_cancer_df}
\title{Veterans Administration Lung Cancer Study}
\format{
A data frame with 137 observations and 8 variables:
\describe{
  \item{trt}{Treatment group (numeric)}
  \item{celltype}{Cell type (factor with 4 levels)}
  \item{time}{Survival time in days (numeric)}
  \item{status}{Censoring status (numeric)}
  \item{karno}{Karnofsky performance score (numeric)}
  \item{diagtime}{Time from diagnosis to randomization (numeric)}
  \item{age}{Age in years (numeric)}
  \item{prior}{Number of prior therapies (numeric)}
}
}
\source{
Data taken from the survival package version 3.8-3
}
\usage{
data(veterans_lung_cancer_df)
}
\description{
This dataset, veterans_lung_cancer_df, is a data frame containing information from a randomized trial
of two treatment regimens for lung cancer. This is a standard survival analysis data set.
}
\details{
The dataset name has been kept as 'veterans_lung_cancer_df' to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the PulmoDataSets package. The suffix 'df' indicates
that this is a data frame object. The original content has not been modified
in any way.
}
