% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{whooping_cough_dk_df}
\alias{whooping_cough_dk_df}
\title{Copenhagen Whooping Cough 1900-1937}
\format{
A data frame with 1,982 weekly observations and 8 variables:
\describe{
  \item{date}{Date of observation (factor)}
  \item{births}{Number of births (integer)}
  \item{day}{Day of month (integer)}
  \item{month}{Month (integer 1-12)}
  \item{year}{Year (integer 1900-1937)}
  \item{cases}{Number of whooping cough cases (integer)}
  \item{deaths}{Number of whooping cough deaths (integer)}
  \item{popsize}{Population size (numeric)}
}
}
\source{
Data taken from the epimdr package version 0.6-5
}
\usage{
data(whooping_cough_dk_df)
}
\description{
This dataset, whooping_cough_dk_df, is a data frame containing weekly incidence data
of whooping cough in Copenhagen, Denmark between January 1900 and December 1937.
It includes 1,982 weekly observations across 8 demographic and epidemiological variables.
}
\details{
The dataset name has been kept as 'whooping_cough_dk_df' to avoid confusion with other
datasets in the R ecosystem. This naming convention helps distinguish this dataset
as part of the PulmoDataSets package. The suffix 'df' indicates that this is a
standard data frame. The original content has not been modified in any way.
}
