% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{whooping_cough_phila_df}
\alias{whooping_cough_phila_df}
\title{Philadelphia Whooping Cough 1925-1947}
\format{
A data frame with 1,200 weekly observations and 5 variables:
\describe{
  \item{YEAR}{Year of observation (integer)}
  \item{WEEK}{Week number (integer)}
  \item{PHILADELPHIA}{Weekly incidence count of whooping cough cases (integer)}
  \item{TIME}{Time index (numeric)}
  \item{TM}{Time marker (integer)}
}
}
\source{
Data taken from the epimdr package version 0.6-5
}
\usage{
data(whooping_cough_phila_df)
}
\description{
This dataset, whooping_cough_phila_df, is a data frame containing weekly incidence
data of whooping cough in Philadelphia between 1925 and 1947, with 1,200 weekly
observations across 5 variables.
}
\details{
The dataset name has been kept as 'whooping_cough_phila_df' to avoid confusion with other
datasets in the R ecosystem. This naming convention helps distinguish this dataset
as part of the PulmoDataSets package. The suffix 'df' indicates that this is a
standard data frame. The original content has not been modified in any way.
}
