% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{whooping_cough_ts}
\alias{whooping_cough_ts}
\title{Whooping Cough Deaths in London (1740-1881)}
\format{
A multivariate time series with 142 annual observations from 1740 to 1881
and 3 variables:
\describe{
  \item{wcough}{Number of whooping cough deaths (integer)}
  \item{ratio}{Death ratio (numeric)}
  \item{alldeaths}{Total deaths from all causes (integer)}
}
}
\source{
Data taken from the DAAG package version 1.25.6
}
\usage{
data(whooping_cough_ts)
}
\description{
This dataset, whooping_cough_ts, is a time series object containing annual counts
of deaths from whooping cough in London from 1740 to 1881, with three measurement
variables recorded each year.
}
\details{
The dataset name has been kept as 'whooping_cough_ts' to avoid confusion with other
datasets in the R ecosystem. This naming convention helps distinguish this dataset
as part of the PulmoDataSets package. The suffix 'ts' indicates that this is a
time series object. The original content has not been modified in any way.
}
